chessr-next/maia-wrapper/requirements.txt
onnxruntime>=1.17.0
websockets>=12.0
python-chess>=1.10.0
numpy>=1.24.0
requests>=2.31.0
pywebview>=5.0
